funciones de API: Close #94
